// frontend/src/api/generated/models/outbound-slip-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipLineDetail } from './outbound-slip-line-detail';
// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipStatus } from './outbound-slip-status';
// May contain unused imports in some cases
// @ts-ignore
import type { OutboundSlipType } from './outbound-slip-type';

/**
 * 
 * @export
 * @interface OutboundSlipDetail
 */
export interface OutboundSlipDetail {
    /**
     * 出荷伝票ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'id'?: number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'slipNumber'?: string;
    /**
     * 
     * @type {OutboundSlipType}
     * @memberof OutboundSlipDetail
     */
    'slipType'?: OutboundSlipType;
    /**
     * 振替伝票番号
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'transferSlipNumber'?: string | null;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'warehouseId'?: number;
    /**
     * 倉庫コード
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'warehouseCode'?: string;
    /**
     * 倉庫名
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'warehouseName'?: string;
    /**
     * 出荷先取引先ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'partnerId'?: number | null;
    /**
     * 取引先コード
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'partnerCode'?: string | null;
    /**
     * 取引先名
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'partnerName'?: string | null;
    /**
     * 出荷予定日
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'plannedDate'?: string;
    /**
     * 配送業者名
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'carrier'?: string | null;
    /**
     * 送り状番号
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'trackingNumber'?: string | null;
    /**
     * 
     * @type {OutboundSlipStatus}
     * @memberof OutboundSlipDetail
     */
    'status'?: OutboundSlipStatus;
    /**
     * 備考
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'note'?: string | null;
    /**
     * 出荷日時
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'shippedAt'?: string | null;
    /**
     * 出荷実行者ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'shippedBy'?: number | null;
    /**
     * キャンセル日時
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'cancelledAt'?: string | null;
    /**
     * キャンセル実行者ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'cancelledBy'?: number | null;
    /**
     * 
     * @type {Array<OutboundSlipLineDetail>}
     * @memberof OutboundSlipDetail
     */
    'lines'?: Array<OutboundSlipLineDetail>;
    /**
     * 作成日時
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'createdAt'?: string;
    /**
     * 更新日時
     * @type {string}
     * @memberof OutboundSlipDetail
     */
    'updatedAt'?: string;
    /**
     * 作成者ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'createdBy'?: number;
    /**
     * 更新者ID
     * @type {number}
     * @memberof OutboundSlipDetail
     */
    'updatedBy'?: number;
}